ai-professor-vs/client/requirements.txt
faster-whisper
pyaudio
numpy
scipy
requests
pygame
gtts
python-dotenv
